Extract graph mouse action handlers into helpers/history.js
