+++ web/src/components/countryhistorymixgraph.js
@@ -4,11 +4,17 @@ import { max as d3Max } from 'd3-array';
 import { connect } from 'react-redux';
 import { forEach } from 'lodash';
 
+import { dispatchApplication } from '../store';
 import formatting from '../helpers/formatting';
 import { getCo2Scale } from '../helpers/scales';
 import { modeOrder, modeColor } from '../helpers/constants';
-import { getExchangeKeys } from '../helpers/zones';
-import { dispatchApplication } from '../store';
+import {
+  getExchangeKeys,
+  createGraphMouseMoveHandler,
+  createGraphMouseOutHandler,
+  createGraphLayerMouseMoveHandler,
+  createGraphLayerMouseOutHandler,
+} from '../helpers/history';
 
 import AreaGraph from './graph/areagraph';
 
@@ -94,29 +100,6 @@ const prepareGraphData = (historyData, colorBlindModeEnabled, displayByEmissions
   };
 };
 
-const getMouseMoveHandler = () => (timeIndex) => {
-  dispatchApplication('selectedZoneTimeIndex', timeIndex);
-};
-const getMouseOutHandler = () => () => {
-  dispatchApplication('selectedZoneTimeIndex', null);
-};
-const getLayerMouseMoveHandler = (setSelectedLayerIndex, isMobile) => (timeIndex, layerIndex, layer, ev, svgRef) => {
-  // If in mobile mode, put the tooltip to the top of the screen for
-  // readability, otherwise float it depending on the cursor position.
-  const tooltipPosition = !isMobile
-    ? { x: ev.clientX - 7, y: svgRef.current.getBoundingClientRect().top - 7 }
-    : { x: 0, y: 0 };
-  setSelectedLayerIndex(layerIndex);
-  dispatchApplication('tooltipPosition', tooltipPosition);
-  dispatchApplication('tooltipZoneData', layer.datapoints[timeIndex].data._countryData);
-  dispatchApplication('tooltipDisplayMode', layer.key);
-  dispatchApplication('selectedZoneTimeIndex', timeIndex);
-};
-const getLayerMouseOutHandler = setSelectedLayerIndex => () => {
-  setSelectedLayerIndex(null);
-  dispatchApplication('tooltipDisplayMode', null);
-};
-
 const getCurrentTime = state =>
   state.application.customDate || (state.data.grid || {}).datetime;
 
@@ -164,14 +147,14 @@ const CountryHistoryMixGraph = ({
   );
 
   // Mouse action handlers
-  const mouseMoveHandler = useMemo(getMouseMoveHandler, []);
-  const mouseOutHandler = useMemo(getMouseOutHandler, []);
+  const mouseMoveHandler = useMemo(createGraphMouseMoveHandler, []);
+  const mouseOutHandler = useMemo(createGraphMouseOutHandler, []);
   const layerMouseMoveHandler = useMemo(
-    () => getLayerMouseMoveHandler(setSelectedLayerIndex, isMobile),
-    [setSelectedLayerIndex, isMobile]
+    () => createGraphLayerMouseMoveHandler(isMobile, setSelectedLayerIndex),
+    [isMobile, setSelectedLayerIndex]
   );
   const layerMouseOutHandler = useMemo(
-    () => getLayerMouseOutHandler(setSelectedLayerIndex),
+    () => createGraphLayerMouseOutHandler(setSelectedLayerIndex),
     [setSelectedLayerIndex]
   );
