+++ web/src/components/countryhistorypricesgraph.js
@@ -6,8 +6,14 @@ import { scaleLinear } from 'd3-scale';
 import { connect } from 'react-redux';
 import { forEach, first } from 'lodash';
 
-import formatting from '../helpers/formatting';
 import { dispatchApplication } from '../store';
+import formatting from '../helpers/formatting';
+import {
+  createGraphMouseMoveHandler,
+  createGraphMouseOutHandler,
+  createGraphLayerMouseMoveHandler,
+  createGraphLayerMouseOutHandler,
+} from '../helpers/history';
 
 import AreaGraph from './graph/areagraph';
 
@@ -38,27 +44,6 @@ const prepareGraphData = (historyData, colorBlindModeEnabled, electricityMixMode
     focusFill,
     valueAxisLabel,
   };
-};
-
-const getMouseMoveHandler = () => (timeIndex) => {
-  dispatchApplication('selectedZoneTimeIndex', timeIndex);
-};
-const getMouseOutHandler = () => () => {
-  dispatchApplication('selectedZoneTimeIndex', null);
-};
-const getLayerMouseMoveHandler = isMobile => (timeIndex, layerIndex, layer, ev, svgRef) => {
-  // If in mobile mode, put the tooltip to the top of the screen for
-  // readability, otherwise float it depending on the cursor position.
-  const tooltipPosition = !isMobile
-    ? { x: ev.clientX - 7, y: svgRef.current.getBoundingClientRect().top - 7 }
-    : { x: 0, y: 0 };
-  dispatchApplication('tooltipPosition', tooltipPosition);
-  dispatchApplication('tooltipZoneData', layer.datapoints[timeIndex].data._countryData);
-  dispatchApplication('tooltipDisplayMode', layer.key);
-  dispatchApplication('selectedZoneTimeIndex', timeIndex);
-};
-const getLayerMouseOutHandler = () => () => {
-  dispatchApplication('tooltipDisplayMode', null);
 };
 
 const getCurrentTime = state =>
@@ -106,10 +91,10 @@ const CountryHistoryPricesGraph = ({
   );
 
   // Mouse action handlers
-  const mouseMoveHandler = useMemo(getMouseMoveHandler, []);
-  const mouseOutHandler = useMemo(getMouseOutHandler, []);
-  const layerMouseMoveHandler = useMemo(() => getLayerMouseMoveHandler(isMobile), [isMobile]);
-  const layerMouseOutHandler = useMemo(getLayerMouseOutHandler, []);
+  const mouseMoveHandler = useMemo(createGraphMouseMoveHandler, []);
+  const mouseOutHandler = useMemo(createGraphMouseOutHandler, []);
+  const layerMouseMoveHandler = useMemo(() => createGraphLayerMouseMoveHandler(isMobile), [isMobile]);
+  const layerMouseOutHandler = useMemo(createGraphLayerMouseOutHandler, []);
 
   return (
     <AreaGraph
