+++ web/src/main.js
@@ -27,7 +27,7 @@ import * as LoadingService from './services/loadingservice';
 import thirdPartyServices from './services/thirdparty';
 
 // Utils
-import { getExchangeKeys } from './helpers/zones';
+import { getExchangeKeys } from './helpers/history';
 import { getCurrentZoneData } from './helpers/redux';
 import { getCo2Scale } from './helpers/scales';
